[Bug] Fix Siubaak/sval#28
