+++ .gitignore
@@ -65,3 +65,5 @@ typings/
 
 # next.js build output
 .next
+
+.vscode
